Change license from Unlicense to CC0
The Unlicense license has a number of problems. Of the first and very
superficial, it specifically talks about software, which is irrelevant
in a repository only containing textual documents.

It additionally has several issues in the way it’s constructed (ie., it
isn’t made by lawyers) and isn’t really a sound “public domain” release.

Rick Moen discusses the downfalls of Unlicense as a viable license at
length in his response to its application for adoption as an OSI
approved OSS license:
https://web.archive.org/web/20170301020915/https://lists.opensource.org/pipermail/license-review/2012-January/001386.html

In its defence, some lawyers did approve of it when it was requested
to be adopted as a legacy license last year. See
http://lists.opensource.org/pipermail/license-review_lists.opensource.org/2020-March/004795.html
and follow discussions through April, May, and June to see discussion
for that.

It is my (personal, not-a-lawyer but license-and-IP-interested layman)
opinion that CC0 is a better license to declare public domain
attribution with than Unlicense and similar "not a license" licenses.

This (Markdown) version of the license was copied directly from
https://github.com/altair-tv/balanced-employee-ip-agreement